Minimum Spanning Tree using Dijkstra's Algorithm
This program is similar to the Minimum Spanning Tree, and was also done
for a course in Data Structures and Algorithms. This program implements
Dijkstra's algorithm on a list of cities in order to find the shortest
path between a starting and ending point. The program takes in your
source, destination, and the maximum radius two cities can be from
eachother in order to be a valid link. The output is a short list from
destination to source of the shortest path between the two nodes given
the radius constraint.